pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-search-benchmark"
version = "0.1.0"
description = "A framework for evaluating code retrieval agents using Git commit history"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Code Search Benchmark Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "GitPython>=3.1.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "pytest>=7.4.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]
semantic = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]

[project.scripts]
benchmark = "benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["benchmark*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = []
